refactor: promote switchyard_trace to nemo_gym as shared infra
The module is harness-neutral by design (depends only on Switchyard's
record schema and the shared Responses converter), so it lives in the
core library where any agent server can reuse it for Switchyard trace
collection; its tests move to tests/unit_tests accordingly.
